backend/requirements.txt
google-generativeai
requests
graphviz
flask 
python-dotenv
pydantic 
flask-cors

